refactor(blog): decompose BlogPageClient into context + components
Extract state into BlogPageContextProvider, split UI into BlogBackground,
BlogFilters, and BlogFeed. Reorganize blog route into _components/ and _lib/.
