+++ apps/blog/src/app/blog/_components/BlogBackground.tsx
@@ -1,0 +1,9 @@
+import { DotGridBackground } from '@abbottland/fui-components';
+
+export default function BlogBackground() {
+  return (
+    <div className="bg-neutral-800 fixed top-0 left-0 w-full h-full border">
+      <DotGridBackground color="default" />
+    </div>
+  );
+}
