+++ apps/blog/src/app/blog/_lib/blogPosts.ts
@@ -1,0 +1,4 @@
+// Re-export types and utilities from the centralized blog module
+export type { BlogPost, BlogCategory } from '../../../types/blog';
+export { categories } from '../../../types/blog';
+export { getAllBlogPosts, getAllBlogCategories } from '../../../lib/blog';
